Add Player Skill Controller
It includes a controller, UI, scriptable objects for skills. Skills can be added to the skill bar from a panel and are activated by the keys 1, 2,..., 9, 0. They have a mana cost and a cooldown time editable in the scriptable object (along with other characteristics).

It also contains the TextMeshPro package (that's where most of the files come from)